Integrate redux-form
