+++ app/javascript/react/screens/App/Overview/screens/MappingWizard/components/MappingWizardClustersStep/MappingWizardClustersStep.js
@@ -1,32 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { noop } from 'patternfly-react';
+import { Field, reduxForm } from 'redux-form';
 
-import DualPaneMapper from '../DualPaneMapper/DualPaneMapper';
-import DualPaneMapperList from '../DualPaneMapper/DualPaneMapperList';
-import DualPaneMapperCount from '../DualPaneMapper/DualPaneMapperCount';
-import DualPaneMapperListItem from '../DualPaneMapper/DualPaneMapperListItem';
-import ClustersStepTreeView from './components/ClustersStepTreeView';
+import ClustersStepForm from './components/ClustersStepForm';
 
 class MappingWizardClustersStep extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      selectedTargetCluster: null, // eslint-disable-line react/no-unused-state
-      selectedSourceClusters: [], // eslint-disable-line react/no-unused-state
-      mappings: [], // eslint-disable-line react/no-unused-state
-      selectedMapping: null // eslint-disable-line react/no-unused-state
-    };
-
-    this.selectSourceCluster = this.selectSourceCluster.bind(this);
-    this.selectTargetCluster = this.selectTargetCluster.bind(this);
-    this.addMapping = this.addMapping.bind(this);
-    this.selectMapping = this.selectMapping.bind(this);
-    this.removeMapping = this.removeMapping.bind(this);
-    this.removeAll = this.removeAll.bind(this);
-  }
-
   componentDidMount() {
     const {
       fetchSourceClustersUrl,
@@ -39,177 +18,30 @@ class MappingWizardClustersStep extends React.Component {
     fetchTargetClustersAction(fetchTargetClustersUrl);
   }
 
-  selectSourceCluster(sourceCluster) {
-    this.setState(prevState => {
-      const isAlreadySelected = prevState.selectedSourceClusters.some(
-        cluster => cluster.id === sourceCluster.id
-      );
-      if (isAlreadySelected) {
-        return {
-          selectedSourceClusters: prevState.selectedSourceClusters.filter(
-            cluster => cluster.id !== sourceCluster.id
-          )
-        };
-      }
-      return {
-        selectedSourceClusters: [
-          ...prevState.selectedSourceClusters,
-          sourceCluster
-        ]
-      };
-    });
-  }
-
-  selectTargetCluster(targetCluster) {
-    this.setState(() => ({ selectedTargetCluster: targetCluster }));
-  }
-
-  addMapping() {
-    const { removeTargetCluster, removeSourceClusters } = this.props;
-    this.setState(prevState => {
-      removeSourceClusters(prevState.selectedSourceClusters);
-      removeTargetCluster(prevState.selectedTargetCluster);
-      return {
-        selectedTargetCluster: null,
-        selectedSourceClusters: [],
-        mappings: [
-          ...prevState.mappings,
-          {
-            ...prevState.selectedTargetCluster,
-            text: prevState.selectedTargetCluster.name,
-            state: {
-              expanded: true
-            },
-            selectable: true,
-            selected: false,
-            nodes: prevState.selectedSourceClusters.map(cluster => ({
-              ...cluster,
-              text: cluster.name,
-              icon: 'fa fa-file-o'
-            }))
-          }
-        ]
-      };
-    });
-  }
-
-  selectMapping(selectedMapping) {
-    this.setState(prevState => ({
-      mappings: prevState.mappings.map(mapping => {
-        if (mapping.id === selectedMapping.id) {
-          return { ...mapping, selected: !mapping.selected };
-        } else if (mapping.id !== selectedMapping.id && mapping.selected) {
-          return { ...mapping, selected: false };
-        }
-        return mapping;
-      }),
-      selectedMapping
-    }));
-  }
-
-  removeMapping() {
-    const { addTargetCluster, addSourceClusters } = this.props;
-    this.setState(prevState => {
-      const { nodes, ...targetCluster } = prevState.selectedMapping;
-      addTargetCluster(targetCluster);
-      addSourceClusters(nodes);
-      return {
-        mappings: prevState.mappings.filter(
-          mapping => !(mapping.id === prevState.selectedMapping.id)
-        ),
-        selectedMapping: null
-      };
-    });
-  }
-
-  removeAll() {
-    const { addTargetCluster, addSourceClusters } = this.props;
-    this.setState(prevState => {
-      prevState.mappings.forEach(mapping => {
-        const { nodes, ...targetCluster } = mapping;
-        addTargetCluster(targetCluster);
-        addSourceClusters(nodes);
-      });
-      return {
-        mappings: []
-      };
-    });
-  }
-
   render() {
     const {
       isFetchingSourceClusters,
       sourceClusters,
       isFetchingTargetClusters,
-      targetClusters
+      targetClusters,
+      removeSourceClusters,
+      removeTargetCluster,
+      addTargetCluster,
+      addSourceClusters
     } = this.props;
-
-    const {
-      selectedTargetCluster,
-      selectedSourceClusters,
-      mappings,
-      selectedMapping
-    } = this.state;
 
     if (!isFetchingSourceClusters && !isFetchingTargetClusters) {
       return (
-        <div className="mapping-wizard-clusters-step">
-          <DualPaneMapper
-            handleButtonClick={this.addMapping}
-            validMapping={
-              !(
-                selectedTargetCluster &&
-                (selectedSourceClusters && selectedSourceClusters.length > 0)
-              )
-            }
-          >
-            {sourceClusters && (
-              <DualPaneMapperList listTitle="Source Clusters">
-                {sourceClusters.map(item => (
-                  <DualPaneMapperListItem
-                    item={item}
-                    key={item.id}
-                    selected={
-                      selectedSourceClusters &&
-                      selectedSourceClusters.some(
-                        sourceCluster => sourceCluster.id === item.id
-                      )
-                    }
-                    handleClick={this.selectSourceCluster}
-                    handleKeyPress={this.selectSourceCluster}
-                  />
-                ))}
-                <DualPaneMapperCount
-                  selectedItems={selectedSourceClusters.length}
-                  totalItems={sourceClusters.length}
-                />
-              </DualPaneMapperList>
-            )}
-            {targetClusters && (
-              <DualPaneMapperList listTitle="Target Clusters">
-                {targetClusters.map(item => (
-                  <DualPaneMapperListItem
-                    item={item}
-                    key={item.id}
-                    selected={
-                      selectedTargetCluster &&
-                      selectedTargetCluster.id === item.id
-                    }
-                    handleClick={this.selectTargetCluster}
-                    handleKeyPress={this.selectTargetCluster}
-                  />
-                ))}
-              </DualPaneMapperList>
-            )}
-          </DualPaneMapper>
-          <ClustersStepTreeView
-            mappings={mappings}
-            selectMapping={this.selectMapping}
-            removeMapping={this.removeMapping}
-            removeAll={this.removeAll}
-            selectedMapping={selectedMapping}
-          />
-        </div>
+        <Field
+          name="clusterMappings"
+          component={ClustersStepForm}
+          sourceClusters={sourceClusters}
+          targetClusters={targetClusters}
+          removeSourceClusters={removeSourceClusters}
+          removeTargetCluster={removeTargetCluster}
+          addTargetCluster={addTargetCluster}
+          addSourceClusters={addSourceClusters}
+        />
       );
     }
     return null;
@@ -243,4 +75,8 @@ MappingWizardClustersStep.defaultProps = {
   addSourceClusters: noop
 };
 
-export default MappingWizardClustersStep;
+export default reduxForm({
+  form: 'mappingWizardClustersStep',
+  initialValues: { clusterMappings: [] },
+  destroyOnUnmount: false
+})(MappingWizardClustersStep);
